pyproject.toml
[tool.poetry]
name = "backendmodular"
version = "0.1.0"
description = "Um projeto demonstrativo de backend com api usando os scripts do Flaskstarter"
authors = ["Felipe Bastos Nunes <felipebastos@ifce.edu.br>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
flaskstarter = "^0.5.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Flask = "^2.0.3"
dynaconf = "^3.1.7"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
